+++ inst/doc/kdtools.R
@@ -17,6 +17,7 @@ kd_sort(y, inplace = TRUE, parallel = TRUE)
 rq = kd_range_query(y, c(0, 0, 0), c(1/4, 1/4, 1/4)); rq
 i = kd_nearest_neighbor(y, c(0, 0, 0)); y[i, ]
 nns = kd_nearest_neighbors(y, c(0, 0, 0), 100); nns
+nni = kd_nn_indices(y, c(0, 0, 0), 10); nni
 
 ## ------------------------------------------------------------------------
 head(tuples_to_matrix(rq))